Rename FlowedMessageUtils to FormatFlowedHelper
