+++ mail/common/src/main/java/com/fsck/k9/mail/internet/MessageExtractor.java
@@ -25,7 +25,7 @@ import timber.log.Timber;
 
 import static com.fsck.k9.mail.internet.CharsetSupport.fixupCharset;
 import static com.fsck.k9.mail.internet.MimeUtility.getHeaderParameter;
-import static com.fsck.k9.mail.internet.FlowedMessageUtils.isFormatFlowed;
+import static com.fsck.k9.mail.internet.FormatFlowedHelper.isFormatFlowed;
 import static com.fsck.k9.mail.internet.MimeUtility.isSameMimeType;
 import static com.fsck.k9.mail.internet.Viewable.Alternative;
 import static com.fsck.k9.mail.internet.Viewable.Html;
@@ -203,7 +203,7 @@ public class MessageExtractor {
             Viewable viewable;
             if (isSameMimeType(mimeType, "text/plain")) {
                 if (isFormatFlowed(part.getContentType())) {
-                    boolean delSp = FlowedMessageUtils.isDelSp(part.getContentType());
+                    boolean delSp = FormatFlowedHelper.isDelSp(part.getContentType());
                     viewable = new Flowed(part, delSp);
                 } else {
                     viewable = new Text(part);
